Moved qa utilities to a separate directory
