+++ swig/grgsm_swig.i
@@ -24,11 +24,11 @@
 #include "grgsm/misc_utils/message_printer.h"
 #include "grgsm/misc_utils/tmsi_dumper.h"
 #include "grgsm/misc_utils/burst_sink.h"
-#include "grgsm/misc_utils/burst_sink_qa.h"
 #include "grgsm/misc_utils/burst_source.h"
-#include "grgsm/misc_utils/burst_source_qa.h"
-#include "grgsm/misc_utils/message_source.h"
-#include "grgsm/misc_utils/message_sink.h"
+#include "grgsm/qa_utils/burst_sink_qa.h"
+#include "grgsm/qa_utils/burst_source_qa.h"
+#include "grgsm/qa_utils/message_source.h"
+#include "grgsm/qa_utils/message_sink.h"
 %}
 
 
@@ -58,10 +58,6 @@ GR_SWIG_BLOCK_MAGIC2(gsm, bursts_printer);
 GR_SWIG_BLOCK_MAGIC2(gsm, burst_sink);
 %include "grgsm/misc_utils/burst_source.h"
 GR_SWIG_BLOCK_MAGIC2(gsm, burst_source);
-%include "grgsm/misc_utils/burst_sink_qa.h"
-GR_SWIG_BLOCK_MAGIC2(gsm, burst_sink_qa);
-%include "grgsm/misc_utils/burst_source_qa.h"
-GR_SWIG_BLOCK_MAGIC2(gsm, burst_source_qa);
 %include "grgsm/misc_utils/extract_system_info.h"
 GR_SWIG_BLOCK_MAGIC2(gsm, extract_system_info);
 %include "grgsm/misc_utils/extract_immediate_assignment.h"
@@ -74,8 +70,13 @@ GR_SWIG_BLOCK_MAGIC2(gsm, controlled_const_source_f);
 GR_SWIG_BLOCK_MAGIC2(gsm, message_printer);
 %include "grgsm/misc_utils/tmsi_dumper.h"
 GR_SWIG_BLOCK_MAGIC2(gsm, tmsi_dumper);
-%include "grgsm/misc_utils/message_source.h"
+
+%include "grgsm/qa_utils/burst_sink_qa.h"
+GR_SWIG_BLOCK_MAGIC2(gsm, burst_sink_qa);
+%include "grgsm/qa_utils/burst_source_qa.h"
+GR_SWIG_BLOCK_MAGIC2(gsm, burst_source_qa);
+%include "grgsm/qa_utils/message_source.h"
 GR_SWIG_BLOCK_MAGIC2(gsm, message_source);
-%include "grgsm/misc_utils/message_sink.h"
+%include "grgsm/qa_utils/message_sink.h"
 GR_SWIG_BLOCK_MAGIC2(gsm, message_sink);
